Separate out components
